Migración de SQLite a MySQL 8.4 + refactor del schema de base de datos (PR 1: Schema Foundation).
Cambios principales
- Base de datos: migración de SQLite a MySQL 8.4. Se reemplazó Microsoft.EntityFrameworkCore.Sqlite por Pomelo.EntityFrameworkCore.MySql y se actualizó EF Core a 8.0.13.
- Schema: 19 tablas con nombres SCREAMING_SNAKE_CASE, columnas decimal(18,2) para montos, y claves foráneas explícitas.
- Nuevas entidades: Categoria, UnidadMedida, Suscripcion, Empresa, Proveedor, Compra, RenglonCompra, Deuda, Pago (reemplaza a PagoVenta).
- Entidades modificadas: Producto (sin stock, nuevo catalogo fields), Venta (sin ID_CAJA), Sucursal (sin NUMERO), Usuario (con SUSCRIPCION_ACTIVA), StockSucursal (PK compuesta, stock decimal), etc.
- Scripts de setup: setup.ps1 que instala MySQL, corre migraciones y seed data automáticamente en una máquina nueva.
- README actualizado con documentación del proyecto, base de datos y cómo levantar.
Cómo probar
1. Clonar el repo y pararse en la raíz
2. Ejecutar .\setup.ps1 (instala MySQL via winget si no está, migra la DB y carga seed data)
3. En una terminal: cd PosWeb && dotnet run (backend en localhost:5196)
4. En otra terminal: cd frontend && npm run dev (frontend en localhost:5173)
5. Loguearse con admin / 123
6. Verificar en Swagger (http://localhost:5196/swagger) que los endpoints respondan
Nota: después de reiniciar la PC, MySQL hay que volver a arrancarlo con .\setup.ps1 (detecta que ya está instalado y solo lo inicia).

Related work items: #12

+++ PosWeb.Domain/Exceptions/StockInvalidoException.cs
@@ -2,9 +2,9 @@
 
 public class StockInvalidoException : DomainException
 {
-    public int Stock { get; }
+    public decimal Stock { get; }
 
-    public StockInvalidoException(int stock)
+    public StockInvalidoException(decimal stock)
         : base($"El stock '{stock}' es inválido")
     {
         Stock = stock;
